requirements.txt
streamlit
python-whois
requests
scikit-learn
pandas
numpy